refactor: reorganize playground components to folder structure
- Move all playground components from flat structure to kebab-case folders
- Add index.ts barrel exports for easier imports (per user preference)
- Create missing SCSS files for ComponentExample and PlaygroundSection
- Update all import statements to use new paths
- Fix theme-toggle and theme-test import paths
- Update SCSS_MIGRATION_PLAN.md to document completed work
