+++ extension/src/playground/components/component-example/index.ts
@@ -1,0 +1,1 @@
+export { ComponentExample } from './component-example';
