+++ extension/src/playground/components/playground-header/index.ts
@@ -1,0 +1,1 @@
+export { PlaygroundHeader } from './playground-header';
