+++ extension/src/playground/components/playground-header/playground-header.scss
