+++ extension/src/playground/components/playground-nav/index.ts
@@ -1,0 +1,1 @@
+export { PlaygroundNav } from './playground-nav';
