+++ extension/src/playground/components/playground-search/index.ts
@@ -1,0 +1,1 @@
+export { PlaygroundSearch } from './playground-search';
