+++ extension/src/playground/components/playground-search/playground-search.scss
