+++ extension/src/playground/components/playground-section/index.ts
@@ -1,0 +1,1 @@
+export { PlaygroundSection } from './playground-section';
