+++ extension/src/playground/components/playground-section/playground-section.scss
@@ -1,0 +1,18 @@
+// Playground Section styles
+.tanaka-playground-section {
+  margin-bottom: calc(var(--mantine-spacing-xl) * 2);
+
+  &__header {
+    margin-bottom: var(--mantine-spacing-lg);
+  }
+
+  &__title {
+    margin-bottom: var(--mantine-spacing-xs);
+  }
+
+  &__content {
+    display: flex;
+    flex-direction: column;
+    gap: var(--mantine-spacing-md);
+  }
+}
