+++ extension/src/playground/components/theme-style-switcher/index.ts
@@ -1,0 +1,1 @@
+export { ThemeStyleSwitcher } from './theme-style-switcher';
