+++ extension/src/playground/components/theme-style-switcher/theme-style-switcher.tsx
