+++ extension/src/playground/components/theme-test/index.ts
@@ -1,0 +1,1 @@
+export { ThemeTest } from './theme-test';
