+++ extension/src/playground/components/theme-test/theme-test.scss
