+++ extension/src/playground/playground-app.tsx
@@ -3,11 +3,11 @@ import '@mantine/core/styles.css';
 import { useDisclosure } from '@mantine/hooks';
 import { useState } from 'react';
 import { withThemeProvider } from '../themes/theme-provider';
-import { ComponentExample } from './components/ComponentExample';
-import { PlaygroundHeader } from './components/PlaygroundHeader';
-import { PlaygroundNav } from './components/PlaygroundNav';
-import { PlaygroundSearch } from './components/PlaygroundSearch';
-import { PlaygroundSection } from './components/PlaygroundSection';
+import { ComponentExample } from './components/component-example';
+import { PlaygroundHeader } from './components/playground-header';
+import { PlaygroundNav } from './components/playground-nav';
+import { PlaygroundSearch } from './components/playground-search';
+import { PlaygroundSection } from './components/playground-section';
 import './playground.scss';
 import './playground-app.scss';
 import { buttonsSection, inputsSection, selectionSection } from './sections';
